Moving search-representative and view-representative
Moving search-representative and view-representative to src/applications/representatives directory. Corrected file path imports and removed invalid href links.
