+++ src/applications/representatives/search-representative/reducers/index.js
@@ -1,5 +1,5 @@
-import formConfig from '../config/form';
 import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
+import formConfig from '../config/form';
 
 import {
   FETCH_REPRESENTATIVE_SEARCH_RESULTS_SUCCESS,
