+++ src/applications/representatives/view-representative/containers/ViewRepresentativeApp.jsx
@@ -4,7 +4,7 @@ import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import FEATURE_FLAG_NAMES from 'platform/utilities/feature-toggles/featureFlagNames';
 import backendServices from 'platform/user/profile/constants/backendServices';
 import { RequiredLoginView } from 'platform/user/authorization/components/RequiredLoginView';
-import { fetchRepresentative } from 'applications/personalization/view-representative/actions';
+import { fetchRepresentative } from 'applications/representatives/view-representative/actions';
 import ViewRepresentativeLayout from '../components/ViewRepresentativeLayout';
 
 function ViewRepresentativeApp(props) {
