+++ src/applications/representatives/view-representative/reducers/index.js
